bugs fixed

+++ api/views.py
@@ -8,6 +8,7 @@ from django.core.files.storage import default_storage
 from PIL import Image
 import os
 from pathlib import Path
+from .utils import handleThumbnail
 
 
 # Create your views here.
@@ -23,10 +24,10 @@ class TasksView(APIView):
 class DeleteTaskView(APIView):
     def delete(self, request, format=None, pk=None):
         task = Task.objects.get(pk=pk)
-        default_storage.delete(task.item_name)
-        thumbnail_name = ''.join(task.item_name.split(
-            '.')[:-1])+'-thumb.'+''.join(task.item_name.split('.')[-1])
-        default_storage.delete(thumbnail_name)
+        if task.item_type == 'file':
+            file, ext = os.path.splitext(task.item_name)
+            default_storage.delete(task.item_name)
+            default_storage.delete(file+'-thumb'+ext)
         task.delete()
         return Response('Task deleted successfully')
 
@@ -111,13 +112,5 @@ class SaveFileView(APIView):
     def post(self, request, format=None):
         file = self.request.FILES['myFile']
         file_name = default_storage.save(file.name, file)
-        BASE_DIR = Path(__file__).resolve().parent.parent
-        print(os.path.join(BASE_DIR, 'media'))
-        thumbnail_name = ''.join(file.name.split(
-            '.')[:-1])+'-thumb.'+''.join(file.name.split('.')[-1])
-        thumb_file = Image.open(file)
-        thumb_file = thumb_file.resize((100, 100), Image.ANTIALIAS)
-        thumb_file.save(os.path.join(BASE_DIR, 'media/'+thumbnail_name),
-                        optimize=True, quality=95)
-        print(thumbnail_name)
+        handleThumbnail(file, file_name, (100, 100))
         return Response(file_name)
